Refactored dev mode API
