+++ .env.example
@@ -67,3 +67,6 @@ DEV_TOOLS_AUTO_OPEN_ENABLED=
 
 # [Optional] Allows to show unpublished content in Help. Disabled by default, set a secret token to enable
 DEV_HELP_PREVIEW_TOKEN=
+
+# [Optional] Enable DEV version of API configuration. Disabled by default, set "1" to enable
+DEV_API_CONFIG=
