+++ libs/generic-view/feature/src/lib/use-dev-views/views/index.ts
@@ -1,0 +1,10 @@
+/**
+ * Copyright (c) Mudita sp. z o.o. All rights reserved.
+ * For licensing, see https://github.com/mudita/mudita-center/blob/master/LICENSE.md
+ */
+
+import contactsView from "./contacts-view"
+
+export default {
+  contacts: contactsView,
+}
